+++ app/templates/analysis.html
@@ -5,6 +5,7 @@
 <script src="http://cdn.leafletjs.com/leaflet/v0.7.7/leaflet.js"></script>
 <script src="http://cdn.jsdelivr.net/npm/jstat@latest/dist/jstat.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/dom-to-image/2.6.0/dom-to-image.min.js"></script>
+<script src="https://cdnjs.cloudflare.com/ajax/libs/FileSaver.js/1.3.8/FileSaver.min.js"></script>
 {% endblock %}
 
 {% block content %}
@@ -34,7 +35,7 @@
     </script>
 
 <script src="{{ url_for('static',filename='js/papaparse.min.js') }}"></script>
-<script type="module" src="{{ url_for('static',filename='js/main.js') }}"></script>
+<script type="module" src="{{ url_for('static',filename='js/analysis.js') }}"></script>
 
 
 {% endblock %}
