+++ app/templates/partials/_upload_modal.html
@@ -56,7 +56,7 @@
                 <table id="field-mapper"></table>
             </div>
             <div class="modal-footer">
-                <button id="upload-file-button" type="button" class="btn btn-primary" onclick="uploadDataset()" disabled>Upload</button>
+                <button id="upload-file-button" type="button" class="btn btn-primary" id="upload-button" disabled>Upload</button>
                 <button id="close-modal-button" type="button" class="btn btn-secondary" data-dismiss="modal">Cancel</button>
             </div>
         </div>
